+++ lib/testcentricity_web/web_core/page_object.rb
@@ -39,9 +39,9 @@ module TestCentricity
     #
     # @param element_hash [Hash] names of buttons (as a symbol) and CSS selectors or XPath expressions that uniquely identifies buttons
     # @example
-    #     buttons new_account_button:  'button#new-account',
-    #             save_button:         'button#save',
-    #             cancel_button:       'button#cancel'
+    #     buttons new_account_button: 'button#new-account',
+    #             save_button:        'button#save',
+    #             cancel_button:      'button#cancel'
     #
     def self.buttons(element_hash)
       element_hash.each(&method(:button))
@@ -63,11 +63,11 @@ module TestCentricity
     #
     # @param element_hash [Hash] names of text fields (as a symbol) and CSS selectors or XPath expressions that uniquely identifies text fields
     # @example
-    #       textfields  name_field:    'input#Name',
-    #                   title_field:   'input#Title',
-    #                   phone_field:   'input#PhoneNumber',
-    #                   fax_field:     'input#FaxNumber',
-    #                   email_field:   'input#Email'
+    #       textfields  name_field:  'input#Name',
+    #                   title_field: 'input#Title',
+    #                   phone_field: 'input#PhoneNumber',
+    #                   fax_field:   'input#FaxNumber',
+    #                   email_field: 'input#Email'
     #
     def self.textfields(element_hash)
       element_hash.each(&method(:textfield))
@@ -112,10 +112,10 @@ module TestCentricity
     #
     # @param element_hash [Hash] names of checkboxes (as a symbol) and CSS selectors or XPath expressions that uniquely identifies checkboxes
     # @example
-    #       checkboxes  hazmat_certified_check:  'input#hazmatCertified',
-    #                   epa_certified_check:     'input#epaCertified',
-    #                   dhs_certified_check:     'input#homelandSecurityCertified',
-    #                   carb_compliant_check:    'input#carbCompliant'
+    #       checkboxes  hazmat_certified_check: 'input#hazmatCertified',
+    #                   epa_certified_check:    'input#epaCertified',
+    #                   dhs_certified_check:    'input#homelandSecurityCertified',
+    #                   carb_compliant_check:   'input#carbCompliant'
     #
     def self.checkboxes(element_hash)
       element_hash.each(&method(:checkbox))